+++ prac_06/guitar_test.py
